Change plumbing of modopts/modreqs:
The previous way of doing it inside the debugger via DkmClrValue was providing unwieldy, so I've now changed things so that modopts and modreqs are plumbed as additional fields in the DkmClrCustomTypeInfo.  This change updates the Roslyn EE accordingly.  We now read the modopts/modreqs of each property or field when doing member expansion and update the DkmClrCustomTypeInfo to plumb these values through.  I also fixed various layers of the formatter to plumb the DkmClrCustomTypeInfo between them (so the MC++ EE can see it) where it wasn't being plumbed previously.

One factor that complicated things considerably was that, in some code paths, value formatting goes through the IDkmClrFormatter interface and the IDkmClrFormatter.GetValueString() that already shipped in VS 2015 RTM does not consume a DkmClrCustomTypeInfo.  To work around this problem without breaking 3rd-party components that use this interface, I added a hook that basically suppresses the call to IDkmClrFormatter when doing MC++, going straight to the internal Roslyn layer, which is more easily changable.

I also updated the debugger assembly references.  This is necessary to consume the new definition of DkmClrCustomTypeInfo.